% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rhello.Rmd.
\name{add.predictor_hybrid}
\alias{add.predictor_hybrid}
\title{Add two predictor_hybrid objects}
\usage{
add.predictor_hybrid(input, other, alpha = 1)
}
\arguments{
\item{input}{A predictor_hybrid object.}

\item{other}{Another predictor_hybrid object.}

\item{alpha}{A numeric scalar multiplier for \code{other}.}
}
\value{
A new predictor_hybrid object representing the sum.
}
\description{
Performs element-wise addition of two predictor_hybrid objects.
Functional predictors are combined using \code{plus.fd()} and \code{times.fd()}.
}