% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{calcMFPCA_predict}
\alias{calcMFPCA_predict}
\title{Prediction of MFPCA Scores}
\usage{
calcMFPCA_predict(
  N,
  p,
  M,
  weights,
  npc,
  uniBasis,
  normFactors_train,
  vectors_train,
  values_train
)
}
\arguments{
\item{N}{Number of observations in the prediction set.}

\item{p}{Number of functional elements.}

\item{M}{Number of multivariate components.}

\item{weights}{Weights for each functional element.}

\item{npc}{Vector containing the number of univariate basis functions.}

\item{uniBasis}{List containing univariate basis decompositions for the prediction data.}

\item{vectors_train}{Eigenvectors derived from training.}

\item{values_train}{Eigenvalues derived from training.}

\item{normFactors}{Normalization factors derived from training.}
}
\value{
A matrix of MFPCA scores for the prediction data.
}
\description{
Calculates Multivariate Functional PCA scores for new (prediction) data based on
the eigenvectors and normalization factors derived from the training data.
}