% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{compute_gram_matrix}
\alias{compute_gram_matrix}
\title{Compute the Gram matrix for a basis object}
\usage{
compute_gram_matrix(basis)
}
\arguments{
\item{basis}{A basis object of class \code{basisfd} from the \code{fda} package.}
}
\value{
A square matrix where entry (i,j) is the inner product of basis functions i and j.
}
\description{
This function calculates the inner product of a basis with itself to form
the Gram matrix.
}