% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_idx_kfold.R
\name{create_idx_kfold}
\alias{create_idx_kfold}
\title{Generate K-Fold Cross-Validation Indices}
\usage{
create_idx_kfold(n_sample, n_fold)
}
\arguments{
\item{n_sample}{Integer. Total number of samples.}

\item{n_fold}{Integer. Number of folds (>= 2).}
}
\value{
A list of length \code{n_fold}; each element has \code{idx_train} and \code{idx_valid}.
}
\description{
Generate K-Fold Cross-Validation Indices
}
\examples{
set.seed(1)
folds <- create_idx_kfold(20, 4)
length(folds)
length(folds[[1]]$idx_valid)

}
