% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{create_idx_train_test}
\alias{create_idx_train_test}
\title{Split Sample Indices into Train and Test Sets}
\usage{
create_idx_train_test(n_sample, train_ratio)
}
\arguments{
\item{n_sample}{Integer. The total number of samples.}

\item{train_ratio}{Numeric. The proportion of samples for the training set (0 to 1).}
}
\value{
A list with named elements \code{idx_train} and \code{idx_test}.
}
\description{
Randomly partitions the indices of a dataset into training and testing sets.
}