% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_normalize_train_test.R
\name{curve_normalize_train_test}
\alias{curve_normalize_train_test}
\title{Normalize Functional Predictors (Train/Test Split)}
\usage{
curve_normalize_train_test(train, test)
}
\arguments{
\item{train}{A \code{predictor_hybrid} object representing the training set.}

\item{test}{A \code{predictor_hybrid} object representing the testing set.}
}
\value{
A list containing:
\describe{
\item{predictor_train}{The normalized training object.}
\item{predictor_test}{The normalized testing object.}
\item{mean_train}{List of mean functions used for centering.}
\item{deno_train}{Vector of scaling factors used.}
}
}
\description{
Standardizes functional predictors to have mean zero and unit integrated variance.
Statistics (mean and SD) are computed from the training set and applied to the test set.
}
\keyword{internal}
