% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{expandBasisFunction}
\alias{expandBasisFunction}
\title{Reconstruct Functions from Scores and Basis}
\usage{
expandBasisFunction(scores, argvals = functions@argvals, functions)
}
\arguments{
\item{scores}{Matrix of scores (coefficients).}

\item{argvals}{Argument values (domain).}

\item{functions}{Basis functions.}
}
\value{
A \code{funData} object containing the reconstructed functions.
}
\description{
Calculates the linear combination of basis functions based on provided scores.
Used to reconstruct the approximated functions.
}