% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{fit.hybridPLS}
\alias{fit.hybridPLS}
\title{Fit Hybrid Penalized PLS with Optional Validation}
\usage{
fit.hybridPLS(W, y, n_iter, lambda, validation_data = NULL)
}
\arguments{
\item{W}{Initial \code{predictor_hybrid} object (Training).}

\item{y}{Numeric vector of response values (Training).}

\item{n_iter}{Integer, number of components to extract.}

\item{lambda}{Numeric vector of smoothing parameters.}

\item{validation_data}{(Optional) A list containing \code{W_test} (predictor_hybrid) and \code{y_test}.}
}
\value{
A list containing the model (rho, xi, beta) and optional \code{validation_rmse}.
}
\description{
Runs the iterative PLS algorithm. If 'validation_data' is provided,
it computes and stores the RMSE on the validation set for each component.
}