% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{fit.mfpca_regression}
\alias{fit.mfpca_regression}
\title{End-to-End MFPCA Regression Pipeline}
\usage{
fit.mfpca_regression(W, y, train_ratio = 0.7, n_pc_max = 10)
}
\arguments{
\item{W}{A \code{predictor_hybrid} object containing all predictors.}

\item{y}{Vector. Response variable.}

\item{train_ratio}{Numeric scalar between 0 and 1 for splitting data (default 0.7).}

\item{n_pc_max}{Integer. Max number of FPCs/PCs to consider (must be <= min(nbasis, n_scalar)).}
}
\value{
List containing the best model fit, optimal component count, and a list of all test RMSEs.
}
\description{
Performs sample splitting, normalization, and dimension reduction via MFPCA (functional)
and PCA (scalar). Selects the optimal number of components by minimizing test set RMSE.
}