% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{fpcaBasis}
\alias{fpcaBasis}
\title{Calculate Functional PCA Basis for Training and Prediction Data}
\usage{
fpcaBasis(
  funDataObject,
  funDataObject_predict,
  nbasis = 10,
  pve = 0.99,
  npc = NULL,
  makePD = FALSE,
  cov.weight.type = "none"
)
}
\arguments{
\item{funDataObject}{The training functional data object.}

\item{funDataObject_predict}{The prediction functional data object.}

\item{nbasis}{Number of basis functions to use for smoothing.}

\item{pve}{Proportion of Variance Explained.}

\item{npc}{Number of Principal Components.}

\item{makePD}{Logical. Enforce positive definiteness of covariance.}

\item{cov.weight.type}{Weighting scheme for covariance estimation.}
}
\value{
A list containing training scores, prediction scores, orthogonality status, functions, and the mean.
}
\description{
A wrapper around the PACE (Principal Analysis by Conditional Expectation) algorithm.
It computes FPCA scores and eigenfunctions for the training data and simultaneously
computes projected scores for a prediction dataset.
}