% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delta.R
\name{get_delta}
\alias{get_delta}
\title{Compute hybrid regression coefficient (delta)}
\usage{
get_delta(W, rho)
}
\arguments{
\item{W}{A \code{predictor_hybrid} object.}

\item{rho}{A numeric vector of PLS scores.}
}
\value{
A single-sample \code{predictor_hybrid} object.
}
\description{
Weighted average of predictor observations with weights \code{rho}.
}
\keyword{internal}
