% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_idx_train}
\alias{get_idx_train}
\title{Generate Train Indices}
\usage{
get_idx_train(n_sample, train_ratio)
}
\arguments{
\item{n_sample}{Integer. Number of samples.}

\item{train_ratio}{Numeric. Ratio of train samples.}
}
\value{
Vector of sorted train indices.
}
\description{
Randomly generates a sorted vector of indices for a training set.
}