% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nu.R
\name{get_nu}
\alias{get_nu}
\title{Compute scalar regression coefficient (nu)}
\usage{
get_nu(y, rho)
}
\arguments{
\item{y}{Response vector.}

\item{rho}{PLS score vector.}
}
\value{
Scalar regression coefficient.
}
\description{
Compute scalar regression coefficient (nu)
}
\keyword{internal}
