% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_penalty_hybrid}
\alias{get_penalty_hybrid}
\title{Construct the roughness penalty matrix for a predictor_hybrid object}
\usage{
get_penalty_hybrid(W)
}
\arguments{
\item{W}{A \code{predictor_hybrid} object.}
}
\value{
A block diagonal penalty matrix of dimension \code{(sum(n_basis_list) + n_scalar)^2}.
}
\description{
Returns a block diagonal matrix representing roughness penalties for each functional predictor
and zeros for the scalar predictors (scalar predictors are not penalized). Each functional component uses its own basis penalty matrix, computed via second derivative penalty via \code{getbasispenalty()} from the \code{fda} package
}