% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_rho}
\alias{get_rho}
\title{Compute PLS Scores}
\usage{
get_rho(d_vec, W)
}
\arguments{
\item{d_vec}{A numeric vector of concatenated coefficients (functional basis coeffs + scalar weights).}

\item{W}{A \code{predictor_hybrid} object containing the data.}
}
\value{
A numeric vector of PLS scores (length n).
}
\description{
Projects the hybrid predictor data W onto the weight direction defined by d_vec
to calculate the PLS scores.
}