% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{inprod.predictor_hybrid}
\alias{inprod.predictor_hybrid}
\title{Algebraic Inner Product for Hybrid Objects}
\usage{
inprod.predictor_hybrid(xi_1, xi_2 = NULL)
}
\description{
Uses pre-computed Gram matrices to calculate inner products
without numerical integration.
}