% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{inprod_pen.predictor_hybrid}
\alias{inprod_pen.predictor_hybrid}
\title{Penalized Inner Product (Algebraic)}
\usage{
inprod_pen.predictor_hybrid(xi_1, xi_2 = NULL, lambda)
}
\description{
Penalized Inner Product (Algebraic)
}