% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{n_sample.fd}
\alias{n_sample.fd}
\title{Get Number of Samples from an fd Object}
\usage{
n_sample.fd(fd_obj)
}
\arguments{
\item{fd_obj}{An object of class \code{fd}.}
}
\value{
Integer representing the number of samples.
}
\description{
Extracts the number of observations (replications) stored in a functional data object.
}