% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{PACE}
\alias{PACE}
\title{PACE Wrapper for funData Objects}
\usage{
PACE(
  funDataObject,
  predData = NULL,
  nbasis = 10,
  pve = 0.99,
  npc = NULL,
  makePD = FALSE,
  cov.weight.type = "none"
)
}
\arguments{
\item{funDataObject}{Training functional data object.}

\item{predData}{Prediction functional data object (optional).}

\item{nbasis}{Number of basis functions.}

\item{pve}{Proportion of variance explained.}

\item{npc}{Number of principal components.}

\item{makePD}{Logical. Enforce positive definiteness.}

\item{cov.weight.type}{Weighting type.}
}
\value{
A list containing the FPCA results formatted as \code{funData} objects.
}
\description{
A wrapper function that applies the PACE algorithm (via \code{.PACE}) to
\code{funData} or \code{irregFunData} objects.
}