% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{PFRModel}
\alias{PFRModel}
\title{PFR Model Handler}
\description{
R6 class to prepare data, fit, and evaluate a Penalized Functional Regression (PFR) model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PFRModel-new}{\code{PFRModel$new()}}
\item \href{#method-PFRModel-get_model_string}{\code{PFRModel$get_model_string()}}
\item \href{#method-PFRModel-fit}{\code{PFRModel$fit()}}
\item \href{#method-PFRModel-computeError}{\code{PFRModel$computeError()}}
\item \href{#method-PFRModel-clone}{\code{PFRModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFRModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-PFRModel-new}{}}}
\subsection{Method \code{new()}}{
Constructs the instance, splits/normalizes data, and formats it for \code{refund::pfr}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFRModel$new(W, y, n_eval, train_ratio)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{W}}{A \code{predictor_hybrid} object containing all predictors.}

\item{\code{y}}{Vector. Response variable.}

\item{\code{n_eval}}{Integer. Number of evaluation points for functional predictors.}

\item{\code{train_ratio}}{Numeric. Split ratio for training data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFRModel-get_model_string"></a>}}
\if{latex}{\out{\hypertarget{method-PFRModel-get_model_string}{}}}
\subsection{Method \code{get_model_string()}}{
Constructs the formula string dynamically based on the selected option.
Uses \code{presmooth = 'fpca.sc'} for functional terms.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFRModel$get_model_string(option = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{option}}{Character. "all", "functional_only", or "scalar_only".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string representing the PFR call.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFRModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-PFRModel-fit}{}}}
\subsection{Method \code{fit()}}{
Evaluates the formula string to fit the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFRModel$fit(option = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{option}}{Character. Model configuration option.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFRModel-computeError"></a>}}
\if{latex}{\out{\hypertarget{method-PFRModel-computeError}{}}}
\subsection{Method \code{computeError()}}{
Predicts on training and testing sets and computes normalized RMSE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFRModel$computeError()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with \code{train_error} and \code{test_error}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PFRModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PFRModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PFRModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}