% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_hybrid_from_coef.R
\name{predictor_hybrid_from_coef}
\alias{predictor_hybrid_from_coef}
\title{Construct a Single-Sample Predictor Hybrid Object from Coefficients}
\usage{
predictor_hybrid_from_coef(format, coef)
}
\arguments{
\item{format}{A \code{predictor_hybrid} template providing structure/basis info.}

\item{coef}{A numeric vector of functional then scalar coefficients.}
}
\value{
A single-sample \code{predictor_hybrid} object.
}
\description{
Construct a Single-Sample Predictor Hybrid Object from Coefficients
}
\keyword{internal}
