% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{replace_obs_hybrid}
\alias{replace_obs_hybrid}
\title{Replace a single observation in a predictor_hybrid object}
\usage{
replace_obs_hybrid(W, i, new_W)
}
\arguments{
\item{W}{A predictor_hybrid object with one or more samples.}

\item{i}{An integer index specifying the observation to replace.}

\item{new_W}{A single-sample predictor_hybrid object to use for replacement.}
}
\value{
A predictor_hybrid object with the i-th observation replaced.
}
\description{
Replaces the i-th observation of a predictor_hybrid object with a new
single-sample predictor_hybrid object.
}