% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualize_y.R
\name{residualize_y}
\alias{residualize_y}
\title{Residualize the response vector}
\usage{
residualize_y(y, rho, nu)
}
\arguments{
\item{y}{Response vector.}

\item{rho}{Score vector.}

\item{nu}{Regression coefficient.}
}
\value{
Residualized response.
}
\description{
Residualize the response vector
}
\keyword{internal}
