% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{scalar_mul.predictor_hybrid}
\alias{scalar_mul.predictor_hybrid}
\title{Multiply a predictor_hybrid object by a scalar}
\usage{
scalar_mul.predictor_hybrid(input, scalar)
}
\arguments{
\item{input}{A \code{predictor_hybrid} object.}

\item{scalar}{A numeric value to multiply all components by.}
}
\value{
A new \code{predictor_hybrid} object scaled by \code{scalar}.
}
\description{
Performs scalar multiplication on both the scalar and functional components
of a \code{predictor_hybrid} object. Functional predictors are scaled using
\code{times.fd()} from the \code{fda} package.
}