% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_normalize_train_test.R
\name{scalar_normalize_train_test}
\alias{scalar_normalize_train_test}
\title{Normalize Scalar Predictors (Train/Test Split)}
\usage{
scalar_normalize_train_test(predictor_train, predictor_test)
}
\arguments{
\item{predictor_train}{A \code{predictor_hybrid} object (train).}

\item{predictor_test}{A \code{predictor_hybrid} object (test).}
}
\value{
A list containing normalized objects and statistics.
}
\description{
Standardizes scalar predictors (Z) to have mean zero and unit variance.
Statistics are computed from the training set.
}
\keyword{internal}
