% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{split.all}
\alias{split.all}
\title{Split Hybrid Predictor and Response Data}
\usage{
\method{split}{all}(W_hybrid, response, train_ratio)
}
\arguments{
\item{W_hybrid}{A \code{predictor_hybrid} object containing all functional and scalar data.}

\item{response}{A numeric vector. The response variable corresponding to the samples in W_hybrid.}

\item{train_ratio}{A numeric scalar between 0 and 1 indicating the proportion of data to use for training.}
}
\value{
A list containing split train and test sets for predictors and response.
}
\description{
Randomly partitions the hybrid predictor object and response vector into
training and testing sets based on a given ratio. Handles any number of
functional predictors.
}