% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{subset_predictor_hybrid}
\alias{subset_predictor_hybrid}
\title{Extract a single observation from a predictor_hybrid object}
\usage{
subset_predictor_hybrid(W, i)
}
\arguments{
\item{W}{A predictor_hybrid object with multiple samples.}

\item{i}{Integer index of the sample to extract.}
}
\value{
A single-sample predictor_hybrid object.
}
\description{
Extract a single observation from a predictor_hybrid object
}