% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{subtr.predictor_hybrid}
\alias{subtr.predictor_hybrid}
\title{Subtract two predictor_hybrid objects}
\usage{
subtr.predictor_hybrid(input, other, alpha = 1)
}
\arguments{
\item{input}{A \code{predictor_hybrid} object.}

\item{other}{Another \code{predictor_hybrid} object to subtract.}

\item{alpha}{A scalar multiplier applied to \code{other} before subtraction (default is 1).}
}
\value{
A new \code{predictor_hybrid} object representing the result of subtraction.
}
\description{
Performs element-wise subtraction of two \code{predictor_hybrid} objects.
Internally uses \code{add.predictor_hybrid(input, other, alpha = -1)}.
}