% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{univExpansion}
\alias{univExpansion}
\title{Calculate Univariate Basis Expansion}
\usage{
univExpansion(
  type,
  scores,
  argvals = ifelse(!is.null(functions), functions@argvals, NULL),
  functions,
  params = NULL
)
}
\arguments{
\item{type}{Decomposition type (e.g., "uFPCA", "splines1D").}

\item{scores}{Matrix of scores.}

\item{argvals}{Argument values.}

\item{functions}{Basis functions.}

\item{params}{Additional parameters.}
}
\value{
A \code{funData} object.
}
\description{
Wrapper function that calls specific basis reconstruction methods (e.g., splines,
FPCA) to reconstruct functions from scores.
}