% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{btwn_normalize_train_test}
\alias{btwn_normalize_train_test}
\title{Inter-Modality Normalization (Between Functional and Scalar)}
\usage{
btwn_normalize_train_test(train, test)
}
\arguments{
\item{train}{A \code{predictor_hybrid} object (train).}

\item{test}{A \code{predictor_hybrid} object (test).}
}
\value{
A list containing normalized objects and the scaling factor.
}
\description{
Scales the scalar predictors (Z) so that their total variability is comparable
to the total variability of the functional predictors. This addresses the
scale invariance issue in PLS by weighting the scalar part.
}