% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{create_idx_kfold}
\alias{create_idx_kfold}
\title{Generate K-Fold Cross-Validation Indices}
\usage{
create_idx_kfold(n_sample, n_fold)
}
\arguments{
\item{n_sample}{Integer. The total number of samples.}

\item{n_fold}{Integer. The number of folds.}
}
\value{
A list of lists, where each sub-list contains \code{idx_train} and \code{idx_valid}.
}
\description{
Creates a list of training and validation indices for K-fold cross-validation.
Uses \code{caret::createFolds} to ensure stratified or random splitting.
}