% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_delta}
\alias{get_delta}
\title{Compute hybrid regression coefficient (delta)}
\usage{
get_delta(W, rho)
}
\arguments{
\item{W}{A \code{predictor_hybrid} object.}

\item{rho}{A numeric vector of PLS scores.}
}
\value{
A single-sample \code{predictor_hybrid} object representing delta.
}
\description{
Calculates the weighted average of the predictor observations W_i,
weighted by the scores rho_i.
}