% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_gram_matrix_block}
\alias{get_gram_matrix_block}
\title{Construct block-diagonal Gram matrix for hybrid predictor}
\usage{
get_gram_matrix_block(obj)
}
\arguments{
\item{obj}{A \code{predictor_hybrid} object.}
}
\value{
A block-diagonal matrix of size \verb{(total_dim × total_dim)} where
functional and scalar components are arranged in order.
}
\description{
Returns a block-diagonal matrix containing the Gram  matrices for
each functional component and an identity matrix for the scalar part.
}