% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_nu}
\alias{get_nu}
\title{Compute scalar regression coefficient (nu)}
\usage{
get_nu(y, rho)
}
\arguments{
\item{y}{A numeric vector of response values.}

\item{rho}{A numeric vector of PLS scores.}
}
\value{
A scalar regression coefficient.
}
\description{
Calculates the projection coefficient of the response vector y onto the score vector rho.
}