% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{get_rho}
\alias{get_rho}
\title{Exact Hybrid Inner Product
Computes <W, xi> using basis inner product matrices.}
\usage{
get_rho(W, xi)
}
\arguments{
\item{W}{The hybrid predictor object (contains coefs for N samples)}

\item{xi}{The hybrid weight object (contains coefs for 1 direction)}
}
\value{
A vector of scores (length N)
}
\description{
Exact Hybrid Inner Product
Computes <W, xi> using basis inner product matrices.
}