% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{predictor_hybrid}
\alias{predictor_hybrid}
\title{Create a predictor_hybrid object}
\usage{
predictor_hybrid(Z, functional_list, eval_point, penalty_order = 2)
}
\arguments{
\item{Z}{Numeric matrix (n_sample x n_scalar).}

\item{functional_list}{List of 'fd' objects.}

\item{eval_point}{Evaluation points.}

\item{penalty_order}{Integer. Derivative order for roughness penalty (default 2).}
}
\description{
Create a predictor_hybrid object
}