% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rhello.Rmd.
\name{rep_fd}
\alias{rep_fd}
\title{Replicate a list of single-sample fd objects multiple times}
\usage{
rep_fd(fd_list, n)
}
\arguments{
\item{fd_list}{A list of \code{fd} objects, each representing a single sample.}

\item{n}{The number of replications desired.}
}
\value{
A list of \code{fd} objects, each with \code{n} replications.
}
\description{
This function broadcasts a list of single-sample functional predictors
by replicating their coefficient columns.
}