% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{residualize_y}
\alias{residualize_y}
\title{Residualize the response vector}
\usage{
residualize_y(y, rho, nu)
}
\arguments{
\item{y}{The current response vector.}

\item{rho}{The current PLS score vector.}

\item{nu}{The regression coefficient.}
}
\value{
The residualized response vector.
}
\description{
Subtracts the projection of y onto rho from y.
}