% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSHybridPLS.Rmd.
\name{scalar_normalize}
\alias{scalar_normalize}
\title{Normalize a scalar predictor matrix}
\usage{
scalar_normalize(scalar_predictor, mean, deno)
}
\arguments{
\item{scalar_predictor}{Numeric matrix.}

\item{mean}{Numeric vector of column means.}

\item{deno}{Numeric vector of column standard deviations.}
}
\value{
Normalized matrix.
}
\description{
Centers columns by 'mean' and scales them by 'deno'.
}